moved encoder code to nand_teensy, deleted teensy_test
