--- other/teensy_test/include/encoder.h
@@ -1,46 +1,0 @@
-/**
- * @file encoder.h
- * @brief Header file for the encoder namespace on the Teensy
- * 
- * This file declares functions in the encoder namespace that create 
- * an interface for the encoder microcontroller and to read encoder data.
- * 
- * @author Sanjay Ravishankar
- * @date 1/31/2026
- * 
- * @author Sanjay Ravishankar
- * @date 2/10/2026 - Added reporting for when packets aren't found
- */
-
-#pragma once
-#include <Arduino.h>
-#define COMM_SERIAL Serial6
-#define COMM_BAUDRATE 115200
-
-namespace encoder
-{
-    /**
-     * @brief Initializes serial communication with the XIAO
-     * Must be called during setup before other encoder functions
-     */
-    void init();
-
-    /**
-     * @brief State machine for encoder communication
-     * Must be called in the loop
-     */
-    void poll();
-
-    /**
-     * @brief Gets the speed as a float
-     * @param[in,out] s: The float pointer for the current front wheel speed in degrees/sec
-     * @return If the operation worked
-     */
-    bool get_speed(float *s);
-
-    /**
-     * @brief Used for making sure packets are being received 
-     * @return How long ago the last packet was received in ms
-     */
-    long lastPacket();
-}
